refactor/cuda: replace multiple instances of get_cuda_backend & get_native_backend
