+++ coaster-nn/src/lib.rs
@@ -39,15 +39,11 @@
 //!     }
 //! }
 //!
+//! use crate::co::frameworks::cuda::get_cuda_backend;
 //! pub fn main() {
 //!     // Initialize a CUDA Backend.
 //!     // Usually you would not use CUDA but let Coaster pick what is available on the machine.
-//!     let framework = Cuda::new();
-//!     let hardwares = framework.hardwares()[0..1].to_vec();
-//!     let backend_config = BackendConfig::new(framework, &hardwares);
-//!     let mut backend = Backend::new(backend_config).unwrap();
-//!     backend.framework.initialise_cublas().unwrap();
-//!     backend.framework.initialise_cudnn().unwrap();
+//!     let backend = get_cuda_backend();
 //!
 //!     // Initialize two SharedTensors.
 //!     let mut x = SharedTensor::<f32>::new(&(1, 1, 3));
